requirements.txt
google-api-python-client
google-genai
gspread
Pillow
python-dotenv
